# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RISE 성과관리 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: 단위과제 목록
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: 단위과제 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: 단위과제 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not Found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: 단위과제 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: 단위과제 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── Indicators ────────────────────────────────────────────
...
  /targets:
    get:
      operationId: listTargets
      tags: [targets]
      summary: 목표값 목록
      parameters:
        - name: indicatorId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IndicatorTarget"
    post:
      operationId: createTarget
      tags: [targets]
      summary: 목표값 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTargetBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorTarget"
  /targets/{id}:
    patch:
      operationId: updateTarget
      tags: [targets]
      summary: 목표값 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTargetBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorTarget"
    delete:
      operationId: deleteTarget
      tags: [targets]
      summary: 목표값 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── Indicator Results ─────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        managerName:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - name
        - status
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        projectId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        managerName:
          type: ["string", "null"]
        status:
          type: string
      required:
        - projectId
        - name
    UpdateTaskBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        managerName:
          type: ["string", "null"]
        status:
          type: string

    # ── Indicators ─────────────────────────────────────────
...
    IndicatorTarget:
      type: object
      properties:
        id:
          type: integer
        indicatorId:
          type: integer
        year:
          type: integer
        targetValue:
          type: ["number", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - indicatorId
        - year
        - createdAt
    CreateTargetBody:
      type: object
      properties:
        indicatorId:
          type: integer
        year:
          type: integer
        targetValue:
          type: ["number", "null"]
        note:
          type: ["string", "null"]
      required:
        - indicatorId
        - year
    UpdateTargetBody:
      type: object
      properties:
        targetValue:
          type: ["number", "null"]
        note:
          type: ["string", "null"]

    # ── Indicator Results ──────────────────────────────────
...